refactor(realworld): split gripper and hand backends
