+++ rlinf/envs/realworld/franka/end_effectors/franka_gripper.py
@@ -12,8 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Compatibility shim for the moved FrankaGripper backend."""
+"""Compatibility shim for the canonical Franka gripper backend."""
 
-from rlinf.envs.realworld.common.end_effectors.franka_gripper import FrankaGripper
+from rlinf.envs.realworld.common.gripper.franka_gripper import FrankaGripper
 
 __all__ = ["FrankaGripper"]
